Update objects.md
Fixed a `,` to a `:`

Something else to consider, when you say:

> This information might include the following facts: he is blue, large and alive

Do you want to start calling attention to attributes? e.g.

> This information might include the following facts: his size is large, his color is blue and his liveliness is healthy

(I admit that last part is a little awkward).

You start talking about attributes of songs later on (length, artist, etc), whereas the whale stuff is just a set of facts.